* remove InputBox * merge key_actions and type_utils to plugin.py
