+++ lib/python/Plugins/Extensions/FileCommander/setup.xml
@@ -1,5 +1,5 @@
 <setupxml>
-	<setup key="filecommander" title="File Commander Settings">
+	<setup key="FileCommander" title="File Commander Settings">
 		<item level="0" text="Add plugin to main menu*" description="Make File Commander accessible from the main menu.">config.plugins.filecommander.add_mainmenu_entry</item>
 		<item level="0" text="Add plugin to Extensions menu*" description="Make File Commander accessible from the Extensions menu.">config.plugins.filecommander.add_extensionmenu_entry</item>
 		<item level="0" text="Save left folder on exit" description="Save the left folder list location on exit.">config.plugins.filecommander.savedir_left</item>
